requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
drf-spectacular==0.27.0
requests==2.31.0